requirements.txt
asgiref==3.7.2
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.2.21
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.8